Restructure the project directory
[~] Restructure the project directory
	move files:
		screen.py
		realm.py
		heaven.py
		paradise.py

	from divine/

	to divine/domains
