+++ divine/__init__.py
@@ -1,6 +1,3 @@
-from .screen import Screen
-from .realm import Realm
-from .heaven import Heaven
-from .paradise import Paradise
+from .domains import Screen, Realm, Heaven, Paradise
 
 __all__ = ['Screen', 'Realm', 'Heaven', 'Paradise']
