+++ divine/domains/__init__.py
@@ -1,0 +1,6 @@
+from .screen import Screen
+from .realm import Realm
+from .heaven import Heaven
+from .paradise import Paradise
+
+__all__ = ['Screen', 'Realm', 'Heaven', 'Paradise']
